Moved rocket_performance_hints_buffer to frontend subscriber
