+++ inc/Engine/Common/PerformanceHints/ServiceProvider.php
@@ -102,10 +102,17 @@ class ServiceProvider extends AbstractServiceProvider {
 				]
 			);
 
+		$this->getContainer()->add( 'config', Config::class )
+			->addArgument( [ 'config_dir_path' => rocket_get_constant( 'WP_ROCKET_CONFIG_PATH' ) ] );
+
+		$this->getContainer()->add( 'tests', Tests::class )
+			->addArgument( $this->getContainer()->get( 'config' ) );
+
 		$this->getContainer()->addShared( 'performance_hints_frontend_subscriber', FrontendSubscriber::class )
 			->addArguments(
 				[
 					$this->getContainer()->get( 'frontend_processor' ),
+					$this->getContainer()->get( 'tests' ),
 				]
 			);
 
@@ -164,13 +171,7 @@ class ServiceProvider extends AbstractServiceProvider {
 				]
 			);
 
-		$this->getContainer()->add( 'config', Config::class )
-			->addArgument( [ 'config_dir_path' => rocket_get_constant( 'WP_ROCKET_CONFIG_PATH' ) ] );
-		$this->getContainer()->add( 'tests', Tests::class )
-			->addArgument( $this->getContainer()->get( 'config' ) );
-
 		$this->getContainer()->addShared( 'performance_hints_warmup_subscriber', WarmUpSubscriber::class )
-			->addArgument( $this->getContainer()->get( 'performance_hints_warmup_controller' ) )
-			->addArgument( $this->getContainer()->get( 'tests' ) );
+			->addArgument( $this->getContainer()->get( 'performance_hints_warmup_controller' ) );
 	}
 }
